Remove v2 folder and put v2 tests into v1 test file
